DF-6367 made work RHI-Cube on MacOS;

+++ Sources/Internal/Render/Material/NMaterial.cpp
@@ -269,7 +269,7 @@ void NMaterial::SetParent(NMaterial *_parent)
         
 
     parent = _parent;
-    sortingKey = (uint32)parent;
+    sortingKey = (uint32)((uint64)parent);
 
     if (parent)
     {
